// src/app/modules/specialists/interface/specialist-test.interface.ts
import {ObjectType} from "../../../shared-modules/types/object.type";

export interface SpecialistTestInterface {
  specialist: SpecialistListsInterface;
}

export interface RejectedOrAcceptedInterface {
  availableSpecialist: string;
  createdAt: string;
  deletedAt: string | number | Date;
  interviewDate: string | number | Date;
  interviewStatus: "accepted" | "rejected";
  interviewTime: string | number;
  isFavorite: boolean;
  newSpecialist: boolean;
  specialistUuid: string;
  updatedAt: string;
  uuid: string;
  vacancyUuid: string;
}
export interface SpecialistListsInterface {
  citizenship: string;
  city: string;
  country: string;
  email: string;
  employment: string;
  image: string;
  image_blured: string;
  languages: "[]";
  languagesFrameworksArray: "[]";
  languagesFrameworksObjects: "{}";
  name: string;
  phone: string;
  surname: string;
  test_answers: QuestionAnswerList[];
  uuid: string;
  found_specialist_for_company_rejected: RejectedOrAcceptedInterface[];
  found_specialist_for_company_succeed: RejectedOrAcceptedInterface[]
}

export interface QuestionAnswerList {
  correctAnswerCount: number;
  createdAt: string;
  foundCompanySpecialistUuid: string;
  interview_test: ObjectType;
  point: number;
  questionAnswerList: QuestionsAnswersInterface;
  specialistUuid: string;
  testUuid: string;
  updatedAt: string;
  uuid: number;
  vacancyUuid: number;
  wrongAnswerCount: number;
}

export interface TestAnswer {
  correctAnswerCount: number;
  createdAt: string;
  foundCompanySpecialistUuid: string;
  interview_test: ObjectType;
  point: number;
  questionAnswerList: ObjectType;
  specialistUuid: string;
  testUuid: string;
  updatedAt: string;
  uuid: number;
  vacancyUuid: number;
  wrongAnswerCount: number;
}
export interface QuestionsAnswersInterface {
  typeOfQuiz?: string;
  questions?: SpecialistAnswers;
  testUuid?: string;
  title?: string;
  point?: number;
  category?: string;
}

export interface SpecialistAnswers {
  answerId: 1;
  answersList:AnswersList[];
  isCorrect: boolean;
  question: string;
  specialistAnswer: string;
}

export interface AnswersList {
  id: number;
  answer: string;
  is_correct: boolean;
}